refactor: split game logic into modules
