+++ index.html
@@ -25,6 +25,6 @@
         <button id="resetButton">Reset Simulation</button>
     </div>
 
-    <script src="js/game.js"></script>
+    <script type="module" src="js/main.js"></script>
 </body>
 </html>
