+++ js/constants.js
@@ -1,0 +1,20 @@
+export const HUMAN_SIZE = 10;
+export const BUSH_SIZE = 12;
+export const TREE_SIZE = 15;
+export const RABBIT_SIZE = 6;
+export const HUMAN_SPEED = 1;
+export const RABBIT_SPEED = 0.7;
+export const MAX_HUNGER = 100;
+export const HUNGER_DECAY_RATE = 0.05;
+export const HUNGER_THRESHOLD_TO_EAT = 50; // Eat if below this and food is available
+export const HUNGER_THRESHOLD_TO_GATHER = 70; // Seek food if below this
+export const GATHER_DISTANCE = 15; // How close to interact
+export const CARRY_CAPACITY = 5; // Max resources a human can carry
+export const BERRIES_PER_BUSH = 5;
+export const WOOD_PER_TREE = 10;
+export const FOOD_PER_RABBIT = 15; // Amount of 'food' gained from hunting a rabbit
+export const HUNGER_PER_BERRY = 20;
+export const HUNGER_PER_FOOD = 25; // Hunger restored by eating 'food' from storage
+export const GATHER_BERRY_DURATION = 60;
+export const CHOP_WOOD_DURATION = 90;
+export const HUNT_RABBIT_DURATION = 75; // Time to 'catch' rabbit when close
